# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Certif",
        "description": "Api de l'appli PENC",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/addforum": {
            "post": {
                "tags": [
                    "Forums"
                ],
                "summary": "Stocke une nouvelle ressource créée",
                "description": "Stocke une nouvelle ressource créée dans le stockage (Il sera utilisé pour stocker dans notre base de données les forums nouvellement créés).",
                "operationId": "fa3b0667189c49d549a2137bf0bc69f5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreForumRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Votre forum a bien été créé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entité non traitable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Il existe déjà un forum pour ce domaine"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/deleteforum/{forum}": {
            "delete": {
                "tags": [
                    "Forums"
                ],
                "summary": "Supprime la ressource spécifiée",
                "description": "Supprime la ressource spécifiée.",
                "operationId": "d9e02131e2b5d7cc5466056a8dd9e9cb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "ID de la ressource à supprimer",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le forum a été supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}